Move component to use it in other tables
